# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BAPS Anand Jamo Ne Jamadu Seva Portal API
  description: Module 3 (Books & Receipts) OpenAPI specification
  version: 3.0.0
  contact:
    name: BAPS Anand Mandir
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
security:
  - CookieAuth: []
  - BearerAuth: []
paths:
...
  /sevaks/search:
    get:
      summary: Search sevaks by code, mobile, or name
      operationId: searchSevaks
      tags:
        - Sevaks
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            enum:
              - 10
              - 20
              - 50
      responses:
        '200':
          description: Paginated search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SevakSearchResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sevaks/export/excel:
    get:
      summary: Download all sevaks as an Excel file
      operationId: exportSevaksExcel
      tags:
        - Sevaks
      responses:
        '200':
          description: Excel file download
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sevaks/export/books-summary:
    get:
      summary: Download Sevak book allocations as an Excel file
      operationId: exportSevakBooksSummary
      tags:
        - Sevaks
      responses:
        '200':
          description: Excel file download
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '429':
          description: Too many requests from this IP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sevaks/{id}:
    put:
      summary: Update a sevak
      operationId: updateSevak
      tags:
        - Sevaks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSevakRequest'
      responses:
        '200':
          description: Sevak updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SevakResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Sevak not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Mobile or WhatsApp already registered to another sevak
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a sevak
      operationId: deleteSevak
      tags:
        - Sevaks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sevak deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Sevak has active donation receipts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Sevak not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UpdateSevakRequest:
      type: object
      required:
        - fullName
        - firstName
        - lastName
        - mobile
        - address
        - mandal
        - kshetra
        - expectedContacts
      properties:
        fullName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        mobile:
          type: string
        altMobile:
          type: string
          nullable: true
        whatsapp:
          type: string
          nullable: true
        address:
          type: string
        mandal:
          type: string
        kshetra:
          type: string
        expectedContacts:
          type: integer
          minimum: 0
    Sevak:
      type: object
      required:
        - id
        - sevakCode
        - fullName
        - firstName
        - lastName
        - mobile
        - address
        - mandal
        - kshetra
        - expectedContacts
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        sevakCode:
          type: string
          description: Format {firstInitial}{lastInitial}2026{zero-padded sequence} (e.g. nb202601)
          example: nb202601
        fullName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        mobile:
          type: string
        altMobile:
          type: string
          nullable: true
        whatsapp:
          type: string
          nullable: true
        address:
          type: string
        mandal:
          type: string
        kshetra:
          type: string
        expectedContacts:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SevakResponse:
      type: object
      required:
        - sevak
      properties:
        sevak:
          $ref: '#/components/schemas/Sevak'
    SevakSearchResponse:
      type: object
      required:
        - sevaks
        - total
        - page
        - limit
      properties:
        sevaks:
          type: array
          items:
            $ref: '#/components/schemas/Sevak'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string